refactor: split ssg route and asset helpers
